camel-twitter: introduce optional truncation
Twitter status updates were truncated to 160 chars if they were longer.

On the one hand, the twitter limit is 140 chars, so this truncation
didn't have any effect. Even after truncation updating twitter would
fail with an exception.

On the other side, there are longer messages that are nevertheless
compliant with the 140 char limit. See for example URLs: every URL is
shortened using twitter's t.co shortener. Thus, even very long URLs
occupy just a handful characters and the character counting, as used in
camel-twitter, fails.

This commit introduces an optional trunctation to 140 chars. That way,
the client can decide whether updates should be truncated or not. There
is a new flag `truncate` that can be used what creating an endpoint:

    twitter://timeline/user?consumerKey=%s&...&truncate=true

However, if deciding for `truncate=false` the client should be aware
that updating twitter may fail raising a `twitter4j.TwitterException`.

I nevertheless set the default value to false, as this best reflects
the previous behaviour and will not break other code.